+++ modules/swagger-jaxrs2/src/test/java/io/swagger/v3/jaxrs2/resources/jspecify/nullmarked/Department.java
@@ -1,6 +1,6 @@
 package io.swagger.v3.jaxrs2.resources.jspecify.nullmarked;
 
-import io.swagger.v3.jaxrs2.resources.jspecify.nullunmarked.Address;
+import io.swagger.v3.jaxrs2.resources.jspecify.nullmarked.nullunmarked.Address;
 
 public class Department {
     private Long id; // default not-null due to @NullMarked on package
