+++ modules/swagger-jaxrs2/src/test/java/io/swagger/v3/jaxrs2/resources/jspecify/nullmarked/nullunmarked/Address.java
@@ -1,4 +1,4 @@
-package io.swagger.v3.jaxrs2.resources.jspecify.nullunmarked;
+package io.swagger.v3.jaxrs2.resources.jspecify.nullmarked.nullunmarked;
 
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
