+++ modules/swagger-jaxrs2/src/test/java/io/swagger/v3/jaxrs2/resources/jspecify/nullmarked/nullunmarked/package-info.java
@@ -1,0 +1,4 @@
+@NullUnmarked
+package io.swagger.v3.jaxrs2.resources.jspecify.nullmarked.nullunmarked;
+
+import org.jspecify.annotations.NullUnmarked;
