--- modules/swagger-jaxrs2/src/test/java/io/swagger/v3/jaxrs2/resources/jspecify/nullunmarked/package-info.java
@@ -1,4 +1,0 @@
-@NullUnmarked
-package io.swagger.v3.jaxrs2.resources.jspecify.nullunmarked;
-
-import org.jspecify.annotations.NullUnmarked;
